Fix hashmap, add hashset, fix pathbuf, add path, fix ion compile
