+++ libstd/src/collections/hash_state.rs
@@ -8,8 +8,7 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-#![unstable(feature = "hashmap_hasher", reason = "hasher stuff is unclear",
-            issue = "27713")]
+
 
 use clone::Clone;
 use default::Default;
@@ -44,19 +43,13 @@ pub struct DefaultState<H>(marker::PhantomData<H>);
 
 impl<H: Default + hash::Hasher> HashState for DefaultState<H> {
     type Hasher = H;
-    fn hasher(&self) -> H {
-        Default::default()
-    }
+    fn hasher(&self) -> H { Default::default() }
 }
 
 impl<H> Clone for DefaultState<H> {
-    fn clone(&self) -> DefaultState<H> {
-        DefaultState(marker::PhantomData)
-    }
+    fn clone(&self) -> DefaultState<H> { DefaultState(marker::PhantomData) }
 }
 
 impl<H> Default for DefaultState<H> {
-    fn default() -> DefaultState<H> {
-        DefaultState(marker::PhantomData)
-    }
+    fn default() -> DefaultState<H> { DefaultState(marker::PhantomData) }
 }
